requirements.txt
numpy==1.21.4
pandas==1.2.4
matplotlib==3.4.3
streamlit==1.2.0
folium==0.12.1.post1
geopandas==0.10.2
Pillow==8.4.0
streamlit_folium==0.4.0
